+++ beacon-chain/node/node.go
@@ -84,12 +84,12 @@ func NewBeaconNode(ctx *cli.Context) (*BeaconNode, error) {
 	// Use custom config values if the --no-custom-config flag is set.
 	if !ctx.GlobalBool(flags.NoCustomConfigFlag.Name) {
 		log.Info("Using custom parameter configuration")
-		if featureconfig.FeatureConfig().DemoConfig {
-			log.Info("Using demo config")
-			params.UseDemoBeaconConfig()
-		} else {
+		if featureconfig.FeatureConfig().MinimalConfig {
 			log.Info("Using minimal config")
 			params.UseMinimalConfig()
+		} else {
+			log.Info("Using demo config")
+			params.UseDemoBeaconConfig()
 		}
 	}
