+++ validator/BUILD.bazel
@@ -7,30 +7,23 @@ load("@io_bazel_rules_docker//contrib:push-all.bzl", "docker_push")
 go_library(
     name = "go_default_library",
     srcs = [
-        "interop.go",
         "main.go",
         "usage.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/validator",
     visibility = ["//validator:__subpackages__"],
     deps = [
-        "//shared/bls:go_default_library",
         "//shared/cmd:go_default_library",
         "//shared/debug:go_default_library",
         "//shared/featureconfig:go_default_library",
-        "//shared/interop:go_default_library",
-        "//shared/keystore:go_default_library",
         "//shared/logutil:go_default_library",
         "//shared/version:go_default_library",
-        "//validator/accounts:go_default_library",
         "//validator/flags:go_default_library",
         "//validator/node:go_default_library",
         "@com_github_joonix_log//:go_default_library",
-        "@com_github_pkg_errors//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_urfave_cli//:go_default_library",
         "@com_github_x_cray_logrus_prefixed_formatter//:go_default_library",
-        "@org_golang_x_crypto//ssh/terminal:go_default_library",
         "@org_uber_go_automaxprocs//:go_default_library",
     ],
 )
@@ -38,7 +31,6 @@ go_library(
 go_image(
     name = "image",
     srcs = [
-        "interop.go",
         "main.go",
         "usage.go",
     ],
@@ -51,23 +43,17 @@ go_image(
     tags = ["manual"],
     visibility = ["//visibility:private"],
     deps = [
-        "//shared/bls:go_default_library",
         "//shared/cmd:go_default_library",
         "//shared/debug:go_default_library",
         "//shared/featureconfig:go_default_library",
-        "//shared/interop:go_default_library",
-        "//shared/keystore:go_default_library",
         "//shared/logutil:go_default_library",
         "//shared/version:go_default_library",
-        "//validator/accounts:go_default_library",
         "//validator/flags:go_default_library",
         "//validator/node:go_default_library",
         "@com_github_joonix_log//:go_default_library",
-        "@com_github_pkg_errors//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_urfave_cli//:go_default_library",
         "@com_github_x_cray_logrus_prefixed_formatter//:go_default_library",
-        "@org_golang_x_crypto//ssh/terminal:go_default_library",
         "@org_uber_go_automaxprocs//:go_default_library",
     ],
 )
