+++ validator/main.go
@@ -1,172 +1,33 @@
 package main
 
 import (
-	"bufio"
-	"encoding/json"
 	"fmt"
-	"io"
-	"io/ioutil"
 	"os"
 	"runtime"
-	"strings"
-	"syscall"
 
 	joonix "github.com/joonix/log"
-	"github.com/pkg/errors"
 	"github.com/prysmaticlabs/prysm/shared/cmd"
 	"github.com/prysmaticlabs/prysm/shared/debug"
 	"github.com/prysmaticlabs/prysm/shared/featureconfig"
 	"github.com/prysmaticlabs/prysm/shared/logutil"
 	"github.com/prysmaticlabs/prysm/shared/version"
-	"github.com/prysmaticlabs/prysm/validator/accounts"
 	"github.com/prysmaticlabs/prysm/validator/flags"
 	"github.com/prysmaticlabs/prysm/validator/node"
 	"github.com/sirupsen/logrus"
 	"github.com/urfave/cli"
 	prefixed "github.com/x-cray/logrus-prefixed-formatter"
 	_ "go.uber.org/automaxprocs"
-	"golang.org/x/crypto/ssh/terminal"
 )
 
 var log = logrus.WithField("prefix", "main")
 
-type unencryptedKeysContainer struct {
-	Keys []*unencryptedKeys `json:"keys"`
-}
-
-type unencryptedKeys struct {
-	ValidatorKey  []byte `json:"validator_key"`
-	WithdrawalKey []byte `json:"withdrawal_key"`
-}
-
 func startNode(ctx *cli.Context) error {
-	// Unsafe start from plain text keys.
-	if unencryptedKeys := ctx.String(flags.UnencryptedKeysFlag.Name); unencryptedKeys != "" {
-		keys, err := loadUnencryptedKeys(unencryptedKeys)
-		if err != nil {
-			return err
-		}
-		validatorClient, err := node.NewValidatorClient(ctx, keys)
-		if err != nil {
-			return err
-		}
-
-		validatorClient.Start()
-		return nil
-	}
-
-	// Interop start from generated keys.
-	if numValidatorKeys := ctx.GlobalUint64(flags.InteropNumValidators.Name); numValidatorKeys > 0 {
-		keys, err := interopValidatorKeys(ctx.GlobalUint64(flags.InteropStartIndex.Name), numValidatorKeys)
-		if err != nil {
-			return err
-		}
-		validatorClient, err := node.NewValidatorClient(ctx, keys)
-		if err != nil {
-			return err
-		}
-
-		validatorClient.Start()
-		return nil
-	}
-
-	// Normal production key start.
-	keystoreDirectory := ctx.String(flags.KeystorePathFlag.Name)
-	keystorePassword := ctx.String(flags.PasswordFlag.Name)
-
-	exists, err := accounts.Exists(keystoreDirectory)
-	if err != nil {
-		log.Fatal(err)
-	}
-	if !exists {
-		// If an account does not exist, we create a new one and start the node.
-		keystoreDirectory, keystorePassword, err = createValidatorAccount(ctx)
-		if err != nil {
-			log.Fatalf("Could not create validator account: %v", err)
-		}
-	} else {
-		if keystorePassword == "" {
-			log.Info("Enter your validator account password:")
-			bytePassword, err := terminal.ReadPassword(int(syscall.Stdin))
-			if err != nil {
-				log.Fatalf("Could not read account password: %v", err)
-			}
-			text := string(bytePassword)
-			keystorePassword = strings.Replace(text, "\n", "", -1)
-		}
-
-		if err := accounts.VerifyAccountNotExists(keystoreDirectory, keystorePassword); err == nil {
-			log.Info("No account found, creating new validator account...")
-		}
-	}
-
-	verbosity := ctx.GlobalString(cmd.VerbosityFlag.Name)
-	level, err := logrus.ParseLevel(verbosity)
+	validatorClient, err := node.NewValidatorClient(ctx)
 	if err != nil {
 		return err
 	}
-	logrus.SetLevel(level)
-
-	validatorKeys, err := accounts.DecryptKeysFromKeystore(keystoreDirectory, keystorePassword)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	validatorClient, err := node.NewValidatorClient(ctx, validatorKeys)
-	if err != nil {
-		return err
-	}
-
 	validatorClient.Start()
 	return nil
-}
-
-func createValidatorAccount(ctx *cli.Context) (string, string, error) {
-	keystoreDirectory := ctx.String(flags.KeystorePathFlag.Name)
-	keystorePassword := ctx.String(flags.PasswordFlag.Name)
-	if keystorePassword == "" {
-		reader := bufio.NewReader(os.Stdin)
-		logrus.Info("Create a new validator account for eth2")
-		log.Info("Enter a password:")
-		bytePassword, err := terminal.ReadPassword(int(syscall.Stdin))
-		if err != nil {
-			log.Fatalf("Could not read account password: %v", err)
-		}
-		text := string(bytePassword)
-		keystorePassword = strings.Replace(text, "\n", "", -1)
-		log.Infof("Keystore path to save your private keys (leave blank for default %s):", keystoreDirectory)
-		text, err = reader.ReadString('\n')
-		if err != nil {
-			log.Fatal(err)
-		}
-		text = strings.Replace(text, "\n", "", -1)
-		if text != "" {
-			keystoreDirectory = text
-		}
-	}
-
-	if err := accounts.NewValidatorAccount(keystoreDirectory, keystorePassword); err != nil {
-		return "", "", errors.Wrapf(err, "could not initialize validator account")
-	}
-	return keystoreDirectory, keystorePassword, nil
-}
-
-func parseUnencryptedKeysFile(r io.Reader) ([][]byte, [][]byte, error) {
-	encoded, err := ioutil.ReadAll(r)
-	if err != nil {
-		return nil, nil, err
-	}
-	var ctnr *unencryptedKeysContainer
-	if err := json.Unmarshal(encoded, &ctnr); err != nil {
-		return nil, nil, err
-	}
-	validatorKeys := make([][]byte, 0)
-	withdrawalKeys := make([][]byte, 0)
-	for _, item := range ctnr.Keys {
-		validatorKeys = append(validatorKeys, item.ValidatorKey)
-		withdrawalKeys = append(withdrawalKeys, item.WithdrawalKey)
-	}
-	return validatorKeys, withdrawalKeys, nil
 }
 
 var appFlags = []cli.Flag{
@@ -225,7 +86,7 @@ contract in order to activate the validator client`,
 						flags.PasswordFlag,
 					},
 					Action: func(ctx *cli.Context) {
-						if keystoreDir, _, err := createValidatorAccount(ctx); err != nil {
+						if keystoreDir, _, err := node.CreateValidatorAccount(ctx); err != nil {
 							log.Fatalf("Could not create validator at path: %s", keystoreDir)
 						}
 					},
